LPS-123900 Abstract Video Preview for use in file picker and item selector url video
